Move mesosScheduler.NewPluginConfig from plugin.go to mesos_scheduler.go
